Makes bot a little smarter. Reveals bot's cards.
Only had to add some lines to the cpuTurn method. Modifies "decider" with the strength of the hand from the bot's point of view. Hidden cards are hidden from the HandCheckBot class by changing the relevant cards to 53, which HandCheckBot will ignore due to its precise design. Cards are changed back, of course. Also makes the bot a bit more likely to fold if the player raises.

Added two lines to reveal the bot's cards after a round.